Add files via upload
Q1. Create a for loop that iterates up to 100 while outputting "fizz" at multiples of 3, "buzz" at multiples of 5 and "fizzbuzz" at multiples of 3 and 5. 
Q2. Destructure the following object 
Q3. Imagine you are going out to do some grocery shopping. 
So you have an array called shoppingList with all the products you want to buy. Now that you are inside of the shop, you have a basket with all the products from your list, but you want to add a few more. 
Create a new array called shoppingBasket, that will be a copy of the shoppingList array, and add some new products into it. 
Q4. Make a Calculator in Javascript which can do operations as Addtion, Subtraction, Multiplication, Division, Square root, Percentage. 
Q5.You are managing a sales department for your company, you want to reward your employees based on the sales made by them during the year. 
The criteria is as follows: 
1. Sales from 0-5000 : 2% 2. Sales from 5001 - 10000 : 5% 3. Sales from 10001 - 20000 : 7% 4. Above 20000 - 10% 
Then log the total commission earned by him. 
Q6. Rewrite the function using '?' or '||' 
Write a loop which prompts for a number greater than 100. If the visitor enters another number – ask them to input again. 
The loop must ask for a number until either the visitor enters a number greater than 100 or cancels the input/enters an empty line. 
Here we can assume that the visitor only inputs numbers. There’s no need to implement a special handling for a non-numeric input in this task. 
Q7.An integer number greater than 1 is called a ​prime​ if it cannot be divided without a remainder by anything except 1 and itself. 
In other words, n > 1 is a prime if it can’t be evenly divided by anything except 1 and n. 
For example, 5 is a prime, because it cannot be divided without a remainder by 2, 3 and 4. 
Write the code which outputs prime numbers in the interval from 2 to n. 
For n = 10 the result will be 2,3,5,7. 
P.S. The code should work for any n, not be hard-tuned for any fixed value 
Q8. Replace Function Expressions with arrow functions in the code below:  

Tip :
Questions in given youtube link

https://www.youtube.com/watch?v=b_5PWhozFJw