Add files via upload
<!-- Question 1) Write a program that will iterate over an array of colors 
and change the background of the pageafter 5 seconds.
-->

<!-- Question 2:Write a program that will display the multiplication table 
of a number on the webpage in a properformat. Take the input from the user.
Input - 5
Output - 5 x 1 = 5   5 x 2 = 10    
and so on..
-->

<!-- Question 3) Make a webpage that will ask the name of the user. 
Then will display a welcome message on the webpage. 
Also the webpage should also have a clock.Using the toggle method of 
classList add a dark mode feature to the website.
-->